Test Scene
Created test scene (duplicate of Main scene)- removed targets, bow, and animated hand (replaced with sphere-hands from original project), disabled the BoworHand script.
Added ControllerTesting script to each hand. Result is when grip button on a controller is pressed, the name of the controller is displayed in the debugging text field.